Moved most of the GitHub classes
Classes related to GitHub are now located in the "OurUmbraco.Community.GitHub" namespace or sub namespaces. Hopefully this should make it easier for others to find the relevant files.
