+++ OurUmbraco.Site/Views/Partials/Home/GitHubContributors.cshtml
@@ -1,4 +1,4 @@
-﻿@model OurUmbraco.Community.Models.GitHubContributorsModel
+﻿@model OurUmbraco.Community.GitHub.Models.GitHubContributorsModel
 
 @if (Model.Contributors.Any())
 {
