+++ OurUmbraco/Community/GitHub/Models/GithubContributorModel.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Runtime.Serialization;
 
-namespace OurUmbraco.Community.Models
+namespace OurUmbraco.Community.GitHub.Models
 {
     [DataContract]
     public class GitHubContributorModel : IGitHubContributorModel
